+++ CHANGELOG.md
@@ -3,8 +3,7 @@ Yii Framework 2 redis extension Change Log
 
 2.1.2 under development
 -----------------------
-
-- no changes in this release.
+- Chg #294: Revert custom retry in `PredisConnection` after update `predis` package (@s1lver)
 
 2.1.1 April 29, 2026
 --------------------